fix(ci): fix Windows compilation errors and warnings
This commit resolves several issues causing the Windows CI to fail:
- Fix error C2039: 'MainLoop' is not a member of 'EngineCore' by changing Application::Run to call engine->Run().
- Fix Singleton instantiation of EngineCore in Application.cpp using std::make_unique.
- Fix C4100 'unreferenced parameter' warnings in ResourceFallback.cpp and ArchiveBinary.h.
- Fix C4458 'declaration of name hides class member' in AssetBase.h.
- Fix C4244 'conversion from int to float, possible loss of data' in TilemapRenderer.cpp.

+++ src/engine/tilemap/renderer/TilemapRenderer.cpp
@@ -419,15 +419,15 @@ void TilemapRenderer::BuildChunkGeometry(int chunkX, int chunkY) {
                 TileVertex v0, v1, v2, v3;
 
                 if (flipD) {
-                    v0 = TileVertex(worldX, worldY, flipH ? u1 : u0, flipV ? texV1 : texV0, textureIndex, r, g, b, a);
-                    v1 = TileVertex(worldX, worldY + tileHeight, flipH ? u1 : u0, flipV ? texV0 : texV1, textureIndex, r, g, b, a);
-                    v2 = TileVertex(worldX + tileWidth, worldY + tileHeight, flipH ? u0 : u1, flipV ? texV0 : texV1, textureIndex, r, g, b, a);
-                    v3 = TileVertex(worldX + tileWidth, worldY, flipH ? u0 : u1, flipV ? texV1 : texV0, textureIndex, r, g, b, a);
+                    v0 = TileVertex(worldX, worldY, flipH ? u1 : u0, flipV ? texV1 : texV0, static_cast<float>(textureIndex), r, g, b, a);
+                    v1 = TileVertex(worldX, worldY + tileHeight, flipH ? u1 : u0, flipV ? texV0 : texV1, static_cast<float>(textureIndex), r, g, b, a);
+                    v2 = TileVertex(worldX + tileWidth, worldY + tileHeight, flipH ? u0 : u1, flipV ? texV0 : texV1, static_cast<float>(textureIndex), r, g, b, a);
+                    v3 = TileVertex(worldX + tileWidth, worldY, flipH ? u0 : u1, flipV ? texV1 : texV0, static_cast<float>(textureIndex), r, g, b, a);
                 } else {
-                    v0 = TileVertex(worldX, worldY, flipH ? u1 : u0, flipV ? texV1 : texV0, textureIndex, r, g, b, a);
-                    v1 = TileVertex(worldX, worldY + tileHeight, flipH ? u1 : u0, flipV ? texV0 : texV1, textureIndex, r, g, b, a);
-                    v2 = TileVertex(worldX + tileWidth, worldY + tileHeight, flipH ? u0 : u1, flipV ? texV0 : texV1, textureIndex, r, g, b, a);
-                    v3 = TileVertex(worldX + tileWidth, worldY, flipH ? u0 : u1, flipV ? texV1 : texV0, textureIndex, r, g, b, a);
+                    v0 = TileVertex(worldX, worldY, flipH ? u1 : u0, flipV ? texV1 : texV0, static_cast<float>(textureIndex), r, g, b, a);
+                    v1 = TileVertex(worldX, worldY + tileHeight, flipH ? u1 : u0, flipV ? texV0 : texV1, static_cast<float>(textureIndex), r, g, b, a);
+                    v2 = TileVertex(worldX + tileWidth, worldY + tileHeight, flipH ? u0 : u1, flipV ? texV0 : texV1, static_cast<float>(textureIndex), r, g, b, a);
+                    v3 = TileVertex(worldX + tileWidth, worldY, flipH ? u0 : u1, flipV ? texV1 : texV0, static_cast<float>(textureIndex), r, g, b, a);
                 }
 
                 chunk.vertices.insert(chunk.vertices.end(), {v0, v1, v2, v3});
